refactor: extract SoundThemes, PercentSettings, CommissionSettings, TipProfileSettings overlays from Dashboard
Extract 4 settings overlay JSX sections (~436 lines) from Dashboard.js into
dedicated components in components/Settings/:
- SoundThemesOverlay.js (167 lines -> component with THEMES array map)
- PercentSettingsOverlay.js (76 lines -> submenu for tip/commission)
- CommissionSettingsOverlay.js (108 lines -> commission config)
- TipProfileSettingsOverlay.js (141 lines -> tip profile selection)

Dashboard.js: 8,434 -> 7,998 lines. Build passes.
